% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrie.R
\name{char_tree}
\alias{char_tree}
\title{Create a new trie}
\usage{
char_tree(words, end = NULL)
}
\arguments{
\item{words}{Vector of words.}

\item{end}{Terminating character.  Default is NULL.}
}
\value{
a trie (list of lists)
}
\description{
Create a trie from a vector of words passed as an argument.
Based on Florent Delmotte's response on http://stackoverflow.com/questions/27060453/how-to-build-an-alphabetical-tree-from-a-list-of-words-in-r.
}
\examples{
char_tree(c('able', 'act', 'acts', 'across',
            'act','bat','babble','bobble'), 'X')

}

