% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrie.R
\name{evaluate_str}
\alias{evaluate_str}
\title{evaluate_str}
\usage{
evaluate_str(str, trie)
}
\arguments{
\item{str}{A character string}

\item{trie}{A trie representing a list of words}
}
\value{
The list evaluated.
}
\description{
evaluate_str
}
\examples{
trie <- char_tree(c('able', 'act', 'acts', 'across','act',
                    'bat','babble','bobble'), 'X')
evaluate_str('acts',trie)

}

