% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrie.R
\name{matching_words}
\alias{matching_words}
\title{Matching words}
\usage{
matching_words(str, trie)
}
\arguments{
\item{str}{A character string that will be compared with the trie.}

\item{trie}{A trie representing a list of words that will be searched.}
}
\value{
A vector of words represented in the trie that match the character string passed in as an argument.
}
\description{
Return a list of words that match a string passed as an argument.
The match is anchored to the beginning of the string, so that words
that start with the string "str" are returned.
}
\examples{
trie <- char_tree(c('able', 'act', 'acts', 'across','act',
                    'bat','babble','bobble'), 'X')
matching_words('act',trie)

}

