% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrie.R
\name{is_word}
\alias{is_word}
\title{Is a word}
\usage{
is_word(str, trie)
}
\arguments{
\item{str}{A character string that may or may not be a word in the trie.}

\item{trie}{The trie to search.}
}
\value{
a Boolean
}
\description{
Verify whether or not a given string is a word in the trie
}
\examples{
trie <- char_tree(c('able', 'act', 'acts', 'across','act',
                    'bat','babble','bobble'), 'X')
is_word('tree', trie)

}

